Реализует выход из формы без сохранения состояния
