+++ src/presenter/point.js
@@ -83,6 +83,7 @@ export default class Point {
   _escKeyDownHandler(evt) {
     if (evt.key === 'Escape' || evt.key === 'Esc') {
       evt.preventDefault();
+      this._pointEditComponent.reset(this._point);
       this._replaceFormToPoint();
     }
   }
@@ -92,6 +93,7 @@ export default class Point {
   }
 
   _handleFormEditClick() {
+    this._pointEditComponent.reset(this._point);
     this._replaceFormToPoint();
   }
