+++ src/view/smart.js
@@ -1,0 +1,41 @@
+import AbstractComponentView from './abstract-component';
+
+export default class Smart extends AbstractComponentView {
+  constructor() {
+    super();
+    this._data = {};
+  }
+
+  updateData(update, noUpdatingElement = false) {
+    if(!update) {
+      return;
+    }
+
+    this._data = Object.assign(
+      {},
+      this._data,
+      update,
+    );
+
+    if (noUpdatingElement) {
+      return;
+    }
+
+    this.updateElement();
+  }
+
+  updateElement() {
+    const prevElement = this.getElement();
+    const parent = prevElement.parentElement;
+    this.removeElement();
+
+    const newElement = this.getElement();
+    parent.replaceChild(newElement, prevElement);
+
+    this.restoreHandlers();
+  }
+
+  restoreHandlers() {
+    throw new Error('Abstract method not implemented');
+  }
+}
